Redo Day 16 with Exponentiation by Squaring
Use method from
https://www.reddit.com/r/adventofcode/comments/7k572l/2017_day_16_solutions/drbqb27/

This method compiles the instructions into a single permutation of
the programs, consisting of the position changes of the index and the
substitutions of each program.

The permutation can be applied to itself to give a new permutation which
represents performing the dance twice (i.e `permutation * permutation =
permutation ^ 2`)

The problem can then be generalized to `program * permutation ^
number_of_dances` and then solved using exponentiation by squaring in
`O(log(n))` time.

So damn clever by @jhawthorn!